% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gComp.R
\name{gComp}
\alias{gComp}
\title{Perform g-computation to estimate difference and ratio effects of an outcome exposure}
\usage{
gComp(
  data,
  formula = NULL,
  Y = NULL,
  X = NULL,
  Z = NULL,
  outcome.type = c("binary", "count", "rate", "continuous"),
  offset = NULL,
  rate.multiplier = 1,
  clusterID = NULL,
  R = 100,
  parallel = c("no", "multicore", "snow"),
  ncpus = getOption("boot.ncpus", 1L),
  subgroup = NULL,
  ...
)
}
\arguments{
\item{data}{the data as a data.frame or tibble containing variables for \code{Y}, \code{X}, and \code{Z} or with variables matching the model variables specified in a user-supplied formula.}

\item{formula}{optional argument which provides the model formula for the \code{glm} function to be used internally. 
Can be supplied as a character or formula object, the function will internally convert it to a formula if not supplied as such. 
If no formula is provided, Y and X must be provided.}

\item{Y}{optional argument which provides the response variable that will be supplied to the \code{glm} function internally.
Must also provide \code{X} in order for the function to work.  Can optinoally provide a formula instead of \code{Y} and \code{X} variables.}

\item{X}{optional argument which provides the binary exposure/treatment group assignment that will be supplied to the \code{glm} function internally.
This variable can be supplied as a factor variable or as a numeric of 0 or 1. Must also provide \code{Y} in order for the function to work.  
Can optinoally provide a formula instead of \code{Y} and \code{X} variables.}

\item{Z}{optional argument which provides the covariates or other variables to adjust for in the \code{glm} function to be used internally.
Can be either a single expression or vector of quoted variable names.}

\item{offset}{An optional numeric value that can be used to specify an *a priori* known component to be included in the linear predictor during fitting.}

\item{rate.multiplier}{An optional numeric value to}

\item{clusterID}{An optinal character string of the variable name to use as the level for resampling if the bootstrap resampling should be done at any level other than random resampling of the dataset.}

\item{R}{The number of bootstrap replicates. Usually this will be a single positive integer.
For importance resampling, some resamples may use one set of weights and others use a different set of weights.  
In this case \code{R} would be a vector of integers where each component gives the number of resamples from each of the rows of weights.}

\item{parallel}{The type of parallel operation to be used (if any) in the underlying bootstrap calculation of the confidence interval.
If missing, the default is taken from the option \code{"boot.parallel"} (and if that is not set, \code{"no"}).}

\item{ncpus}{integer: number of processes to be used in parallel operation for the bootstrap calculation.
Typically one would chose this to the number of available CPUs.}

\item{subgroup}{An optinal character string of the variable name to use for subgroup analyses.}

\item{...}{Other named arguments for \code{boot} which are passed unchanged each time it is called. Arguments to boot should follow the specifications in the \code{\link{boot}} package}

\item{family}{a description of the error distribution and link function to be used in the model when calling the \code{glm} function internally. 
For gComp this can currenlty only be a binomial distribution and the link function will be forced to 'logit' but updates are forthcoming to handle various distributions.}

\item{cl}{An optional \pkg{parallel} or \pkg{snow} cluster for use if \code{parallel = "snow"}.
If not supplied, a cluster on the local machine is created for the duration of the \code{boot} call.}
}
\description{
Obtain a point estimate and 95% confidence limits for the difference and ratio between treatment and non-treatment (or exposed/not exposed) groups.
}
\examples{
## Obtain the risk difference and risk ratio for cardiovascular disease or death between 
## patients with and without diabetes.  
data(cvdd)
diabetes <- gComp(cvdd, formula = "cvd_dth ~ DIABETES + AGE + SEX + BMI + CURSMOKE + PREVHYP", 
outcome.type = "binary", R = 10)

}
\keyword{gComp}
