% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_tibble.R
\name{get_results_tibble}
\alias{get_results_tibble}
\title{Take predicted dataframe and calculate the outcome (risk difference/ratio, incidence rate difference/ratio, marginal difference, and number needed to treat)}
\usage{
get_results_tibble(predict.df, outcome.type, X, rate.multiplier)
}
\arguments{
\item{predict.df}{}

\item{outcome.type}{}

\item{X}{}

\item{rate.multiplier}{}
}
\description{
Take predicted dataframe and calculate the outcome (risk difference/ratio, incidence rate difference/ratio, marginal difference, and number needed to treat)
}
\examples{


}
