% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gComp.R
\name{print.gComp}
\alias{print.gComp}
\title{Print estimates of difference and ratio effects obtained in the bootstrap computations of the g-computation}
\usage{
\method{print}{gComp}(gComp.res)
}
\arguments{
\item{gComp.res}{an object of class\code{gComp} with R bootstrap iterations of a parameter estimate for difference and ratio.}
}
\description{
Print results from bootstrap computations of the g-computation
}
\examples{
## Obtain the risk difference and risk ratio for cardiovascular disease or death 
## between patients with and without diabetes, while controlling for 
## age,
## sex, 
## BMI, 
## whether the individual is currently a smoker, and 
## if they have a history of hypertension. 
data(cvdd)
diabetes.result <- gComp(data = cvdd, Y = "cvd_dth", X = "DIABETES", 
Z = c("AGE", "SEX", "BMI", "CURSMOKE", "PREVHYP"), family = "binomial", R = 10)
print(diabetes.result)


}
\keyword{print.gComp}
