% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gComp.R
\name{gComp}
\alias{gComp}
\title{Perform g-computation to estimate difference and ratio effects and 95\%
confidence intervals for a treatment/exposure.}
\usage{
gComp(
  data,
  outcome.type = c("binary", "count", "rate", "continuous"),
  formula = NULL,
  Y = NULL,
  X = NULL,
  Z = NULL,
  subgroup = NULL,
  offset = NULL,
  rate.multiplier = 1,
  R = 200,
  clusterID = NULL
)
}
\arguments{
\item{data}{(Required) A data.frame or tibble containing variables for
\code{Y}, \code{X}, and \code{Z} or with variables matching the model
variables specified in a user-supplied formula. Data set should also
contain variables for the optinal \code{subgroup} and \code{offset}, if
they are specified}

\item{outcome.type}{(Required) Character argument to describe the outcome
type. Acceptable responses, and the corresponding error distribution and
link function used in the \code{glm}, include: \describe{
\item{binary}{(Default) A binomial distribution with link = 'logit' is
used.} \item{count}{A Poisson distribution with link = 'log' is used.}
\item{rate}{A Poisson distribution with link = 'log' is used.}
\item{continuous}{A gaussian distribution with link = 'identity' is used.}
}}

\item{formula}{(Optional) Default NULL. An object of class \code{formula} (or
one that can be coerced to that class) which provides the the complete
model formula, similar to the formula for the glm function in R (e.g. `Y ~
X + Z1 + Z2 + Z3`). Can be supplied as a character or formula object. If no
formula is provided, Y and X must be provided.}

\item{Y}{(Optional) Default NULL. Character argument which specifies the
outcome variable. Can optionally provide a formula instead of \code{Y} and
\code{X} variables.}

\item{X}{(Optional) Default NULL. Character argument which specifies the
exposure variable (or treatment group assignment), which can be binary,
categorical, or continuous. This variable can be supplied as a factor
variable, a numeric variable coded 0 or 1, or a continuous variable.
Preferrably, \code{X} is supplied as a factor with the lowest level set to
the desired comparator. Numeric variables are accepted, and coerced to
factor with lowest level being the smallest number. Character variables are
not accepted and will throw an error. Can optionally provide a formula
instead of \code{Y} and \code{X} variables.}

\item{Z}{(Optional) Default NULL. List or single character vector which
specifies the names of covariates or other variables to adjust for in the
\code{glm} function to be used internally. Does not allow interaction
terms.}

\item{subgroup}{(Optional) Default NULL. Character argument that indicates
subgroups for stratified analysis. Effects will be reported for each
category of the subgroup variable. Variable will be automatically converted
to a factor if not already.}

\item{offset}{(Optional, only applicable for rate outcomes) Default NULL.
Character argument which specifies the person-time denominator for rate
outcomes to be included as an offset in the Poisson regression model.
Numeric variable should be on the linear scale; function will take natural
log before including in the model.}

\item{rate.multiplier}{(Optional, only applicable for rate outcomes) Default
1. Numeric value to multiply to the rate-based effect measures. This option
facilitates reporting effects with interpretable person-time denominators.
For example, if the person-time variable (offset) is in days, a multiplier
of 365*100 would result in estimates of rate differences per 100
person-years.}

\item{R}{(Optional) Default 200. The number of data resamples to be conducted
to produce the bootstrap confidence interval of the estimate.}

\item{clusterID}{(Optional) Default NULL. Character argument which specifies
the variable name for the unique identifier for clusters. This option
specifies that clustering should be accounted for in the calculation of
confidence intervals. The \code{clusterID} will be used as the level for
resampling in the bootstrap procedure.}
}
\value{
An object of class \code{gComp} which is a list with components:
  \item{summary}{Summary providing parameter estimates and 95\% confidence
  limits of the outcome difference and ratio} \item{results.df}{Data.frame
  with parameter estimates, 2.5\% confidence limit, and 97.5\% confidence
  limit each as a column} \item{n}{Number of observations in the original
  dataset} \item{R}{Number of bootstrap iterations}
  \item{boot.result}{Data.frame containing the results of the \code{R}
  bootstrap iterations of the g-computation} \item{contrast}{Cntrast levels
  compared} \item{family}{Error distribution used in the model}
  \item{formula}{Model formula used to fit the \code{glm}}
  \item{predicted.data}{A tibble with the predicted values for both treatment
  and no treatment counterfactual predicitions for each observation in the
  original dataset}
}
\description{
Obtain a point estimate and 95\% confidence interval for the
  difference and ratio between treatment and non-treatment (or exposed/not
  exposed) groups.
}
\examples{
## Obtain the risk difference and risk ratio for cardiovascular disease or death between
## patients with and without diabetes.
data(cvdd)
diabetes <- gComp(cvdd, formula = "cvd_dth ~ DIABETES + AGE + SEX + BMI + CURSMOKE + PREVHYP",
outcome.type = "binary", R = 200)

}
\keyword{gComp}
