% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predict_df.R
\name{make_predict_df}
\alias{make_predict_df}
\title{Using the \code{glm} results, predict outcomes for each individual at each level of treatment/exposure}
\usage{
make_predict_df(glm.res, df, X, subgroup = NULL, offset = NULL)
}
\arguments{
\item{glm.res}{}

\item{df}{}

\item{X}{}

\item{subgroup}{}

\item{offset}{}
}
\description{
Using the \code{glm} results, predict outcomes for each individual at each level of treatment/exposure
}
\examples{

}
