% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointEstimate.R
\name{pointEstimate}
\alias{pointEstimate}
\title{Perform g-computation to estimate difference and ratio effects of an exposure}
\usage{
pointEstimate(
  data,
  outcome.type = c("binary", "count", "rate", "continuous"),
  formula = NULL,
  Y = NULL,
  X = NULL,
  Z = NULL,
  subgroup = NULL,
  offset = NULL,
  rate.multiplier = 1
)
}
\arguments{
\item{data}{(Required) A data.frame or tibble containing variables for
\code{Y}, \code{X}, and \code{Z} or with variables matching the model
variables specified in a user-supplied formula. Data set should also
contain variables for the optinal \code{subgroup} and \code{offset}, if
they are specified.}

\item{outcome.type}{(Required) Character argument to describe the outcome
type. Acceptable responses, and the corresponding error distribution and
link function used in the \code{glm}, include: \describe{
\item{binary}{(Default) A binomial distribution with link = 'logit' is
used.} \item{count}{A Poisson distribution 
with link = 'log' is used.}
\item{rate}{A Poisson distribution with link = 'log' is used; ideal for 
 events/person-time outcomes.} \item{continuous}{A gaussian distribution 
 with link = 'identity' is used.}
}}

\item{formula}{(Optional) Default NULL. An object of class "formula" (or one
that can be coerced to that class) which provides the the complete model
formula, similar to the formula for the glm function in R (e.g. `Y ~ X + Z1
+ Z2 + Z3`). Can be supplied as a character or formula object. If no
formula is provided, Y and X must be provided.}

\item{Y}{(Optional) Default NULL. Character argument which specifies the
outcome variable. Can optionally provide a formula instead of \code{Y} and
\code{X} variables.}

\item{X}{(Optional) Default NULL. Character argument which specifies the
exposure variable (or treatment group assignment), which can be binary,
categorical, or continuous. This variable can be supplied as a factor
variable, a numeric variable coded 0 or 1, or a continuous variable.
Preferably, \code{X} is supplied as a factor with the lowest level set to
the desired referent. Numeric variables are accepted, and coerced to
factor with lowest level (referent) being the smallest number. Character variables are
not accepted and will throw an error. Can optionally provide a formula
instead of \code{Y} and \code{X} variables.}

\item{Z}{(Optional) Default NULL. List or single character vector which
specifies the names of covariates or other variables to adjust for in the
\code{glm} function. All variables should either be factors, continuous, 
or coded 0/1 (i.e. not character variables). Does not allow interaction terms.}

\item{subgroup}{(Optional) Default NULL. Character argument that indicates
subgroups for stratified analysis. Effects will be reported for each
category of the subgroup variable. Variable will be automatically converted
to a factor if not already.}

\item{offset}{(Optional, only applicable for rate outcomes) Default NULL.
Character argument which specifies the person-time denominator for rate
outcomes to be included as an offset in the Poisson regression model.
Numeric variable should be on the linear scale; function will take natural
log before including in the model.}

\item{rate.multiplier}{(Optional, only applicable for rate outcomes) Default
1. Numeric value to multiply to the rate-based effect measures. This option
facilitates reporting effects with interpretable person-time denominators.
For example, if the person-time variable (offset) is in days, a multiplier
of 365*100 would result in estimates of rate differences per 100
person-years.}
}
\value{
A list containing the following: \itemize{
  \item{parameter.estimates} {Point estimates for the risk difference, risk
  ratio, odds ratio, incidence rate difference, incidence rate ratio, mean
  difference and/or number needed to treat, depending on the outcome.type}
  \item{n} {Number of observations provided to the model} \item{contrast}
  {Contrast levels compared} \item{family} {Error distribution used
  in the model} \item{formula} {Model formula used to fit the
  \code{glm}} \item{Y} {The response variable} \item{covariates}
  {Covariates used in the model} \item{predicted.data} {A tibble with the
  predicted values for the exposed and unexposed counterfactual predictions 
  for each observation in the original dataset} }
}
\description{
Generate a point estimate of the outcome difference and ratio
  using G-computation
}
\details{
The \code{pointEstimate} function executes the following steps on
  the data: 
  \enumerate{ 
  \item Fit a regression of the outcome on the exposure
  and relevant covariates, using the provided data set. 
  \item Using the
  modelfit in step 1, predict counterfactuals (e.g. calculate predicted
  outcomes for each observation in the data set under each level of the
  treatment/exposure). 
  \item Estimate the marginal difference/ratio of
  treatment effect by taking the difference or ratio of the average of all
  observations under the treatment/no treatment regimes. }

  As counterfactual predictions are generated with random sampling of the
  distribution, users should set a seed (\code{\link[base]{set.seed}} for
  reproducible confidence intervals.
}
\examples{
## Obtain the risk difference and risk ratio for cardiovascular disease or death
## between patients with and without diabetes, while controlling for
## age,
## sex,
## BMI,
## whether the individual is currently a smoker, and
## if they have a history of hypertension.
data(cvdd)
ptEstimate <- pointEstimate(data = cvdd, Y = "cvd_dth", X = "DIABETES",
Z = c("AGE", "SEX", "BMI", "CURSMOKE", "PREVHYP"), outcome.type = "binary")

}
\seealso{
\code{\link{gComp}}
}
